+++ init.lua
@@ -1,8 +1,5 @@
 require("configs.settings")
-
-require('packer').startup(function(use)
-  use 'wbthomason/packer.nvim'
-
-  use 'EdenEast/nightfox.nvim'
-  vim.cmd('colorscheme nightfox')
-end)
+require("plugins")
+require("treesitter-config")
+vim.cmd('colorscheme nightfox')
+-- vim.cmd('colorscheme sobrio')