+++ CHANGELOG.md
@@ -6,10 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Instalação do plugin lazy.nvim
+- Instalação do Telescope
+
+## [0.3.0] - 2023-06-08
+- Instalação do Treesitter
+
+## [0.2.0]  - 2023-06-07
+### Added
+- Instalação do plugin Packer
 
 ## [0.1.1] - 2023-05-05
-
 ### Added
 
 - Adição do Changelog.